refactor: rename test file and fix test case
