+++ app/src/tests/App.test.tsx
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom"
 import { test, expect } from "vitest"
 import { render } from "@testing-library/react"
-import App from "./App"
+import App from "../App"
 
 test("demo", () => {
   expect(true).toBe(true)
